Add browser mappings for core (#48)
Adding stub for write-output and browser mappings
