+++ .chronus/changes/main-2025-0-30-16-56-53.md
@@ -1,0 +1,7 @@
+---
+changeKind: fix
+packages:
+  - "@alloy-js/core"
+---
+
+Add browser mappings to enable running in browser
