+++ packages/core/package.json
@@ -25,6 +25,10 @@
       "import": "./dist/src/components/index.js"
     }
   },
+  "browser": {
+    "./dist/src/write-output.js": "./dist/src/write-output.browser.js",
+    "./src/write-output.ts": "./src/write-output.browser.ts"
+  },
   "scripts": {
     "build-src": "babel src -d dist/src --extensions .ts,.tsx",
     "build-testing": "babel testing -d dist/testing --extensions .ts,.tsx",
