+++ packages/core/src/index.ts
@@ -18,4 +18,5 @@ export * from "./name-policy.js";
 export * from "./refkey.js";
 export * from "./render.js";
 export * from "./utils.js";
+export * from "./write-output.js";
 import "./debug.js";
