+++ packages/core/src/write-output.browser.ts
@@ -1,0 +1,3 @@
+export function writeOutput(output: any, basePath: string = "") {
+  console.warn("writeOutput is not supported in a browser environment.");
+}
